Switch to node-sass (from node-sass-chokidar)
It looks like `node-sass-chokidar` doesn't play as nicely with npm
SCSS modules like Bourbon, so we're switching it out in favor of
the base node-sass.

The chokidar version won't work with paths.js, since you have to
pass the include paths as arguments to the script in package.json.

This makes sense, but the main reason for not using the base node-sass
was apparently [a problem with Mac + docker](https://github.com/facebookincubator/create-react-app/issues/1986)
which for better or wose is not at play in my dev environment.